Concatenation integer strings
As with integers and floats, strings in Python can be added, using a process called concatenation, which can be done on any two strings.

Even if your strings contain numbers, they are still added as strings rather than integers.

Adding a string to a number produces an error, as even though they might look similar, they are two different entities.

Strings can also be multiplied by integers. 
This produces a repeated version of the original string. 
The order of the string and the integer doesn't matter, but the string usually comes first.

*************
Strings can't be multiplied by other strings. Strings also can't be multiplied by floats, even if the floats are whole numbers.