Variables 14.1
Variables


A variable lets you store a value by assigning it to a name. The name can be used to refer to the value later in the program.

For example, in game development, you would use a variable to store the points of the player. What’s the point of playing if you don’t know who’s winning right!

To assign a variable, use one equals sign